SNAP-682 Removed the QtSerialization and tld dependencies. They are not needed here.
Removed references to the watermark functions since it was deprecated (removed from Cassandra itself, actually!)
Removed references to the DowngradingConsistency retry policy. It will be removed from Cassandra.
Added the CASS_VALUE_TYPE_DURATION type in the parseValue() switch(). But I did not attempt to implement it at this time.